2-practice-grpc/messenger/client-py/requirements.txt
grpcio==1.40.0
grpcio-tools==1.40.0
protobuf==3.20